app/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chatbot-service"
version = "0.1.0"
description = "Secure FastAPI gateway for Vertex AI chatbot deployments"
authors = [{ name = "Platform Team", email = "platform@example.com" }]
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110,<0.112",
  "uvicorn[standard]>=0.24,<0.30",
  "pydantic>=2.5,<3.0",
  "pydantic-settings>=2.2,<3.0",
  "python-dotenv>=1.0,<2.0",
  "google-cloud-aiplatform>=1.38,<2.0",
  "google-cloud-logging>=3.6,<4.0",
  "google-auth>=2.23,<3.0",
  "httpx>=0.25,<0.28"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4,<8.0",
  "pytest-asyncio>=0.21,<0.24",
  "pytest-cov>=4.1,<5.0",
  "ruff>=0.3.0,<0.6.0",
  "mypy>=1.6,<1.9"
]

[tool.pytest.ini_options]
pythonpath = "src"
addopts = "-q"

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B", "UP", "ASYNC", "S"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
warn_unused_ignores = true
ignore_missing_imports = true
namespace_packages = true
mypy_path = "src"
